+++ viking.js
@@ -1358,12 +1358,27 @@ Viking.Coercions.String = {
         return value;
     }
 };
+
+
+
 // Viking.View
 // -----------
 //
 // Viking.View is a framework fro handling view template lookup and rendering.
 // It provides view helpers that assisst when building HTML forms and more.
-Viking.View = {};
+Viking.View = Backbone.View.extend({}, {
+    
+    // Override the original extend function to support merging events
+    extend: function(protoProps, staticProps) {
+        
+        if (protoProps  && protoProps.events) {
+            _.defaults(protoProps.events, this.prototype.events);
+        }
+        
+        return Backbone.View.extend.call(this, protoProps, staticProps);
+      }
+});
+
 Viking.View.Helpers = {};
 
 
@@ -3347,8 +3362,6 @@ Viking.View.Helpers.mailTo = function (email, name, options) {
 
     return Viking.View.Helpers.contentTag('a', name, options, false);
 };
-
-